refactor(crypto): CRP-2597 move MasterPublicKeyId protobuf from registry/crypto to types (#2406)
Moves the `MasterPublicKeyId` Protocol Buffers message definition, and
subsequently also the contained `EcdsaCurve`, `EcdsaKeyId`,
`SchnorrAlgorithm`, `SchnorrKeyId`, `VetKdCurve`, `VetKdKeyId`
definitions from the `registry.crypto.v1` package to the `types.v1`
package, without making any changes to the types in the process.

This step is necessary so that we can later add an `optional
MasterPublicKeyId key_id = 6;` field to the `types.v1.NiDkgId` message.
Without the move, this would create a circular dependency between
`registry/crypto/v1/crypto.proto` and `types/v1/types.proto` because
`registry/crypto/v1.crypto.proto` imports `types/v1/types.proto` (so as
to use `types.v1.SubnetId` in the `ChainKeySigningSubnetList`);

Given that all message definitions remain unchanged and we only change
the package, the _wire_ format of the messages is unchanged, which
should make this change safe. However, changes like this lead to the
`//pre-commit:buf-breaking` test to fail (even though `buf.yaml` is
configured to use WIRE), which is a known deficiancy in the underlying
`buf` compatibility-check library:
https://github.com/bufbuild/buf/issues/2318. Because of this, we set the
`CI_OVERRIDE_BUF_BREAKING` flag to override this.