refactor(cketh/ckerc20): use consistently `LedgerSubaccount` (#2487)
([XC-221](https://dfinity.atlassian.net/browse/XC-221)):
`LedgerSubaccount` was introduced with #2151 to be used for events
stored in stable memory and was used in particular for events related to
deposits. Unfortunately, withdrawal events were already prepared to
received subaccounts, but they used a different type with a different,
less efficient, encoding. This PR removes the old type to only keep
`LedgerSubaccount`, which is in preparation of supporting subaccounts
for withdrawals.

Remarks:
1. There are already instances of `ReceivedEthEvent` and
`ReceivedErc20Event` with some `LedgerSubaccount` value on the
ckSepoliaETH minter.
2. All other events `EthWithdrawalRequest`, `Erc20WithdrawalRequest`,
and `ReimbursementRequest` that used the old type for subaccount only
had instances on the ckETH and ckSepoliaETH minter that had `None` for
value. This allows to use `LedgerSubaccount` for those events in a
backwards-compatible manner.


[XC-221]:
https://dfinity.atlassian.net/browse/XC-221?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ