fix(ckbtc): Fix a problem with retrieve_btc_min_amount setting (#2435)
The `retrieve_btc_min_amount` setting was both used as a configuration
and as a dynamically calculated value based on recent fees. This led to
a problem that changing this setting through canister upgrades would be
ineffective because it would be overwritten by the calculated value.

The fix is to introduce another `fee_based_retrieve_btc_min_amount` in
the state to hold the calculated value, and let
`retrieve_btc_min_amount` always retain its initial setting.