feat: add function Topology::get_subnet to PocketIC library (#2561)
This PR adds the function `Topology::get_subnet` to the PocketIC library
to get the subnet to which a canister belongs independently of whether
the canister exists. Such a function is needed to migrate StateMachine
tests to PocketIC since StateMachine test API exposes subnet IDs
publicly independently of whether canisters are already deployed.