feat(nns): Changed the default "voting power refreshed" timestamp to Sep 1. (#2570)
More precisely, midnight UTC.

# Background/Motivation

The value used to be Nov 5 (one second after midnight). Thus, this value
has been decreased by a couple of months or so.

Originally, I set it to Nov 5, because I figured that would be the
earliest feasible date we could release the
`Neuron.voting_power_refreshed_timestamp_seconds field`. But then, it
was decided that legacy neurons should be "pre-aged" by a couple of
months.

IIUC, the rationale for pre-aging is that it has been known for a long
time that this feature would be implemented. Also, many neurons probably
haven't taken any actions that would refresh voting power in a long
time. Therefore, we want them to start losing voting power sooner than 6
months (but still leave plenty of time for them to act to maintain full
potential voting power).

# References

This is part of the [periodic confirmation feature][feature].

[feature]: https://dashboard.internetcomputer.org/proposal/132411

Closes https://dfinity.atlassian.net/browse/NNS1-3432.