test: CON-1434 Prepare `cup_compatibility_test` for removal of `pb::TaggedNiDkgTranscript` (#2537)
Currently, NiDkg transcripts are stored in the summary using
`pb::TaggedNiDkgTranscript`, which contains the `pb::NiDkgTranscript`
and the NiDkg tag. The latter data is redundant, as the NiDkg tag is
also already part of `pb::NiDkgTranscript`. Therefore we should be able
to replace `pb::TaggedNiDkgTranscript` with `pb::NiDkgTranscript`
entirely.

To do this, we first need to tell the `cup_compatibility_test` that the
current replica version only creates instances of
`pb::TaggedNiDkgTranscript`, where `tagged_transcript.tag ==
tagged_transcript.transcript.tag`. Otherwise, the test would continue to
create instances where this isn't the case. These instances of
`pb::TaggedNiDkgTranscript` would not be correctly deserialized as
`pb::NiDkgTranscript` once we make the switch in a future version.

This is achieved by implementing the `HasId<NiDkgTag>` trait for
`NiDkgTranscript`.