feat: CON-675 Reduce acceptable finalization-certification gap to one (#2565)
The recent change in https://github.com/dfinity/ic/pull/2450 behaves as
expected, so we further reduce the acceptable gap between latest
certified and finalized heights to one round. This means that the gap
should not grow bigger than two, as long as execution rounds are shorter
than four seconds.