feat(ckbtc): Use the new KYT canister in ckbtc withdrawal flow (#2240)
Add the new KYT canister as a dependency of the ckBTC minter, and use it
instead of the old KYT canister for address checking during ckBTC -> BTC
withdrawal.

Because the check_address call does not require payment, calling
retrieve_btc on the ckBTC miner will no longer charge a KYT fee.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>