perf: EXC-1793: Add criterion load benchmark (#2587)
The benchmark installs the specified number of `load_simulator`
canisters using the `canister_creator` and executes the specified amount
of rounds. It measures the total time taken to complete the rounds and
the throughput, i.e., the number of rounds executed per second
(essentially, the FR).

By default, each `load_simulator` canister runs a periodic timer with a
one-second interval and accesses stable memory every fifth call.

This benchmark is useful for debugging and benchmarking scheduler and
sandbox eviction changes. For more realistic testnet load tests, refer
to the `dfinity/subnet-load-tester` project.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>