test(ckbtc): replay events (#2564)
([XC-231](https://dfinity.atlassian.net/browse/XC-231)): Ensure that
minter events can be replayed from current state, similarly to what was
done for the ckETH minter in dd8247fc8417de7c18f38a7ca6ff2b9d0371bebc.
The minter state is modified through events which are stored in stable
memory and can be retrieve via the debug endpoint `get_events`. Since
the state itself is not saved in stable memory, the minter upon upgrades
will replay all events to get back to an equivalent state.

This PR ensures that the current state always allow to replay those
events and that the resulting state is the expected one. That means:
1. By default such a test uses a local checked-in copy of the minter
events and can be run offline. We for example control that the final
balances are the expected ones.
2. A refresh of the stored minter events can be manually triggered to
retrieve the events from the minter.
3. The events are replayed by default without checking invariants every
time since this would take took long (roughly 1h). For convenience, a
manual test was added to be able to replay invariants while re-checking
invariants at every step.

[XC-231]:
https://dfinity.atlassian.net/browse/XC-231?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>