chore: Add log_and_halt_installation_on_error sleep (#2605)
When we get installation screen recordings from node providers, the
error log often comes so fast that the previous logs don't show up on
the console for a single frame. Adding a sleep helps ensure that the
logs show up on the console before the error log comes and wipes it.