feat(nns): Increase the neurons limit to 380K (#2603)
The number of neurons keeps growing. Some analysis showed that we would
be able to increase the limit to 420K. However, given that we are
actively working on moving all neurons to stable storage, we would like
to increase the limit in a more conservative way. The 380K limit should
give us ~18 weeks.