feat: [MR-592] Define and enforce callback limits (#1838)
Define a subnet-wide soft cap on callbacks; and a per-canister
guaranteed callback quota; and enforce them when outgoing requests are
being enqueued.

The best-effort call model means that canisters can make arbitrarily
many calls. Too many best-effort calls will likely result in timeouts
and "compact reject responses" (one single `u64`), but also one
`Callback` and one `CallContext` plus overhead. Which are all kept
in-memory at all times.

In order to limit the resulting memory usage, we enforce a subnet-wide
cap (initially set to 1M callbacks). We no not want reaching this cap to
result in canisters being unable to make any calls at all, so we also
introduce a per-canister guaranteed callback quota (initially 50
callbacks). For canisters, this means virtually no practical limit under
all but the most extreme conditions. And in the eventuality of a DoS
attack, a guaranteed quota of 50 concurrent calls each.

The marginal cost of one call is in the hundreds of bytes, so there is
no need for a hard cap. A soft cap is also simpler to implement, e.g. by
avoiding artificial limitations, such as being forced to divide the
available pool among execution threads. A soft cap may be temporarily
exceeded (e.g. if very close to the limit and canisters on two or more
execution threads each make significant numbers of calls), but that is
OK because canisters can always rely on their guaranteed quota.

In terms of implementation, we calculate the subnet-wide "usage" at the
beginning of every round; adjust it down every time a call is made; and
use the maximum of that and the remaining canister quota as the limit
for individual canisters (in effect, they can use the shared pool if not
full; and fall back on their quota if consumed).