chore(CON-1412): Avoid handling `VetKdKeyId`s in the IDKG component (#2388)
This is a PR to secure, that the IDKG protocol can only be invoked with
IDKG compatible variants of the `MasterPublicKeyId`.

It does so by introducing a new type `IDkgMasterPublicKeyId`, which
wraps `MasterPublicKeyId` and assures, that only an IDKG compatible
variant is uses (Either `::Schnorr` or `ECDSA`).
Then it changes the type at some places in the code and then makes
changes upstream, such that the Key only needs to be checked once, and
from then on it is guaranteed to be IDKG compatible key.

The newly introduced `::VetKD` variant necessitates the introduction of
this change. The VetKeys will be using the NiDKG protocol for its key
generation, and therefore, a lot of the functions in the
`consensus::idkg` submodule don't make sense to be used with the new
variant.

The keys get transformed back into `MasterPublicKeyId` for
serialization, so this change does not introduce compatibility issues.
When deserializing a `IDkgMasterPublicKeyId`, we need to check that it
is IDKG compatible and reject otherwise. Since we did not introduce
VetKeys yet (and in fact lack the code to do so), there can be no
`::VetKD` variants of `MasterPublicKeyId` yet in the state.