feat(crypto): CRP-2597 Extend NiDkgTag with HighThresholdForKey variant (#2445)
Extends the `NiDkgTag` with a new variant
`NiDkgTag::HighThresholdForKey` that holds a `MasterPublicKeyId`.

### Notes: 
* Makes the representation for `NiDkgTag` explicit with `#[repr(isize)]`
because a `#[repr(inttype)]` must be provided on an enum if it has a
non-unit variant with a discriminant
([E0732](https://doc.rust-lang.org/error_codes/E0732.html)). The
representation should be the same as it was before, given that `isize`
is [the default representation for
enums](https://doc.rust-lang.org/reference/items/enumerations.html#discriminants).

Note that casting `NiDkgTag` with `as`, e.g., to `i32` or even `isize`
is no longer possible because "an `as` expression can be used to convert
enum types to numeric types only if the enum type is unit-only or
field-less" (see [Enumeration
casting](https://doc.rust-lang.org/reference/items/enumerations.html#casting)
and [E0605](https://doc.rust-lang.org/error_codes/E0605.html)).
* The conversion `impl TryFrom<i32> for NiDkgTag` had to be removed,
because it is no longer possible, because an `i32` itself is no longer
sufficient for instantiating an `NiDkgTag`.

This conversion was used in two places: (1) `impl TryFrom<NiDkgIdProto>
for NiDkgId` and (2) `build_tagged_transcripts_map`. There, the
conversion is now implemented directly.
* No longer Implements `EnumIter` for `NiDkgTag` because this would
require a `Default` implementation for `MasterPublicKeyId`, which does
not make sense, at least in production. Instead implements `EnumCount`.