feat(nns): Voting rewards are now in proportion to total *potential* voting power. (#2385)
Whereas, before, we did not have a distinction between **potential**
voting power vs. **deciding** voting power. When neurons refresh
themselves, there is no difference between these. Whereas, when a neuron
does not refresh itself recently enough, the amount of voting power in
its ballots (its "deciding" voting power) is less than its potential
voting power.

# Example

Suppose there are two neurons:

* A has 100 potential voting power, and refreshed recently, and as a
result, its deciding voting power is also 100.
* B has 1000 potential voting power, but has not refreshed in a long
time, and as a result, it's deciding voting power is just 500.

Then, the total potential voting power is 1100. If both neurons vote,
then A gets 100 / 1100 of the reward purse, while B gets only 500 / 1100
of the reward purse, not 1000 / 1100.

You might also intuitively guess that the neurons get 100 / 600 and 500
/ 600 of the reward purse, but this is (also) incorrect. The reason for
this design is to keep the amount of rewards to refreshed neurons the
same as before, not more. In other words, it should not be that some
other neuron goes to sleep, and as a result, you get more rewards. It
should only be that the sleeper neuron gets less.

# References

This is part of the "periodic confirmation" feature that was recently
approved in proposal [132411][prop].

[prop]: https://dashboard.internetcomputer.org/proposal/132411

Closes https://dfinity.atlassian.net/browse/NNS1-3411.

[👈 Previous PR][prev] | [Next PR (tangent) 👉][next]

[prev]: https://github.com/dfinity/ic/pull/2375
[next]: https://github.com/dfinity/ic/pull/2470