feat(crypto): CRP-2597 Use NiDkg-specific MasterPublicKeyId in NiDkgTag (#2620)
Introduces a `MasterPublicKeyId` enum dedicated for NI-DKG, called
`NiDkgMasterPublicKeyId`, and uses that in
`NiDkgTag::HighThresholdForKey` instead of the `MasterPublicKeyId`, so
as to be explicit via the type system on which types of keys can be used
with NiDkg (and which cannot). With this, we no longer need to match on
`MasterPublicKeyId::Ecdsa` or `MasterPublicKeyId::Schnorr` in the
context of NI-DKG.

To be fully future-proof, the respective protobuf messages remain
unchanged, however, i.e., the new `key_id` fields in the `NiDkgId` and
the `TaggedNiDkgTranscript` keep type `MasterPublicKeyId`.