feat: EXC-1787: Penalize canisters at the end of the round (#2601)
The logic of penalizing canisters for full execution and distributing
the priority across the subnet has been moved to the end of the round.
This change is necessary for priority-based eviction.
    
Newly created canisters join the first execution round with an
accumulated priority of zero, which breaks two existing tests, but is
not a practical issue. To compensate for this, during the reset round,
we set the accumulated priority to the canister's compute allocation
instead of the default zero priority.

There are also a few minor optimizations.