feat: Restrict legacy ICQC code to one subnet (#2647)
The legacy code for ICQC (inter-canister query calls), aka predecessor
of composite queries, is still enabled in system and verified
application subnets. There's only one actual user of this legacy feature
and it's Distrikt. The new team that took over is going to migrate off
of this legacy feature but some extra time is required. When this is
done we can completely remove this legacy feature.

On the IC side, given the recent load issues we experienced on mainnet,
we would like to open up some of the (currently) unused verified
application subnets. However, we wouldn't want to open the door to this
legacy feature (only composite queries should be used). To satisfy all
requirements, the proposed change in this PR is to restrict the legacy
feature only on Distrikt's subnet until the team can complete the
migration.