feat: Evict sandboxed based on scheduler priorities (#2590)
When the canister is executed, it is most likely its Sandbox will not be
evicted from the cache, since we are keeping Sandboxes based using LRU
logic. At the same time, the Scheduler decreases its priority so it is
less likely that it will be executed. So our caching of Sandbox
processes is almost the worst possible.

Solution:
Propagate Scheduler priorities for where we evict Sandbox processes and
do evictions based on the lowest priority. That change should decrease
the number of cache misses.

Results:
This change (along with accompanying changes) improves the cache hit
rate from 1% to 75-80% when having 7000 active canisters in the subnet.

---------

Co-authored-by: Andriy Berestovskyy <andriy.berestovskyy@dfinity.org>