feat(crypto): CRP-2596 create VetKdProtocol trait (#2569)
Introduce a `VetKdProtocol` trait and implement it for
`CryptoReturningOk`.

The trait is designed so that underneath we can implement the
cryptographic protocol specified in the π<sub>vetbls-agg2</sub> scheme
specified in section 5.3 of the [vetKeys research
paper](https://eprint.iacr.org/2023/616.pdf), in its _optimistic_
variant, where the meaning of optimistic is described in section 9 of
the research paper. The π<sub>vetbls-agg2</sub> scheme is implemented in
[ic-crypto-internal-bls12-381-vetkd](https://github.com/dfinity/ic/blob/master/rs/crypto/internal/crypto_lib/bls12_381/vetkd/src/lib.rs).