feat: CRP-2641 CRP-2615 Add vetKeys related management canister endpoints (#2633)
This PR defines names and arguments of the new functions:

- ReshareChainKey
- VetKdPublicKey
- VetKdEncryptedKey

A corresponding interface spec PR can be found
[here](https://github.com/dfinity/portal/pull/3763). For now, all
functions remain unimplemented.

In addition, we implement routing of these methods by calling
`route_idkg_message` with the correct `MasterPublicKeyId`. This function
will be renamed and tested for the vet KD case in a follow up PR.