fix(ckbtc): ensure tasks are always rescheduled (#2630)
The tasks `ProcessLogic` and `RefreshFeePercentiles` are executed on a
timer at different frequencies. Since #2435, the second task is also
guarded by a `TimerLogicGuard`, which is the same type of guard used by
`ProcessLogic`, meaning it could be the case that a task of type
`RefreshFeePercentiles` prevents a task of type `ProcessLogic` from
being executed, which is particularly the case upon canister
initialization or upgrade as all tasks are scheduled at the same time.

This is problematic because the re-scheduling of a task of type
`ProcessLogic` only happened _after_ having tried to obtain a
`TimerLogicGuard`, meaning that in case this failed, due to for example
the running of a `RefreshFeePercentiles`, the task `ProcessLogic` would
not be rescheduled.

The fix simply consists of having the rescheduling of each task as a
scope guard that is declared _before_ trying to retrieve the concurrency
guard for the given task type. To be able to unit test this, some
refactoring was needed to abstract away some canister specific methods
related to time management to be able to mock those.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>