feat(IDX): speed up dockerfile rebuilds
This moves less-often updated dependencies to the beginning of the
Dockerfile and more-often updated dependencies to the end of the
Dockerfile to ensure as many layers are possible can be reused when
dependencies like e.g. Rust and Bazel are updated. Other dependencies
like `yq` or `motd` are rarely updated.

This also removes the `RUST_VERSION` docker arg to instead read from the
`rust-toolchain.toml`.

Some `COPY` directives are also moved towards the end of the Dockerfile
because they are so fast to run.