test(sns): Add test that all SNS canisters can be upgraded using the advance-target-version mechanism (#2722)
The test is simple - we deploy an SNS, add upgrade steps for all 6
canisters, then advance the target version, then check that the SNS
eventually gets to the desired version.

This tests that advance-target-version's upgrade logic is capable of
working for all 6 canisters, at least in a normal happy scenario.