chore(fuzzing): Allow `no_op_logger` in StateMachine (#2736)
Currently, the StateMachine sets `log_level: None` by default and
switches it to `Warning` in `fn replica_logger`, which is
counterintuitive. The PR addresses it by setting the value default to
`Some(Level::Warning)` and if `None` is passed, `no_op_logger` is used.

This is required for fuzzing, since we maintain a StateMachine
invocation in a global state and would need to reduce it's memory
footprint as much as possible for fuzzing long hours.