chore(IDX): clarify launch-remote-vm script (#2745)
This replaces the `genrule` with an `sh_binary` for which bazel sets the
environment variables previously baked into the `genrule`. This makes
the script's dependencies slightly more declarative and means the script
does not have to `cd` into a specific directory to run, and instead can
use the variables directly.

The script can also now be written without having to escape symbols.