feat(ckbtc): re-evaluate discarded UTXOs (#2674)
Deposited UTXOs of bitcoin can be discarded and not be minted to ckBTC
for one of two reasons:

1. Either their value is too small to cover the KYT fee;
2. Or, the KYT check marked them as being tainted.

With the upcoming improvements regarding how UTXOs will be checked, both
the KYT fee and the decision of whether a UTXO is tainted will change,
requiring that already discarded UTXOs are re-evaluated.

Remarks:
- This re-evaluation will be user-triggered, by calling (again)
`update_balance` with the appropriate ledger account (principal and
optionally subaccount). It cannot be done automatically, e.g., with a
timer, because the ledger account that was given when the UTXOs was
initially discarded was not recorded.
- Re-evaluation of discarded UTXOs come with some subtleties, e.g., no
event should be added if the outcome is unchanged, avoid deduplication
of UTXOs in the minter's state, that need to be tested. To that end, the
method `update_balance` was refactored to allow unit testing (mock any
IC-specific method).
- Future work:
- Add a ledger account to the discarded UTXOs to minimize the number of
discarded UTXOs that need to be re-evaluate for a given account.
- Add a timestamp regarding the last time a discarded UTXO was
evaluated, to ensure that discarded UTXO are re-evaluated at most once a
day.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>