feat: CRP-2611 Introduce new context type for VetKD requests (#2629)
We extend the existing `SignWithThresholdContext` by adding another
`ThresholdArguments` variant of type `VetKdArguments`. Contexts of this
type cannot be created yet.

In addition, the compiler forces us to define a cycles cost for vet KD
requests. For now we choose the same cost as for tEcdsa and tSchnorr,
but this may be refined further at a later stage.

There are some parts in consensus where we only want to handle tSchnorr
and tEcdsa contexts (as they rely on IDKG), but ignore vet KD contexts
(as they rely on NiDKG). For that reason we add a new type
`IDkgSignWithThresholdContext` which wraps the reference to a
`SignWithThresholdContext`, but only if that context may be handled by
IDKG. Functions in the IDKG component are then changed to accept the new
type.