docs(governance): More streamlined procedure for generating and submitting proposals. (#2637)
In particular, the instructions are now more amenable to copy n' paste.
At the top of each chunk of commands, there are some variables that the
user (i.e. release tsar) fills in. Then, the rest of the commands in the
chunk can be run **_without modification._**

# Future Work

Now that no manual work is needed to generate a complete proposal
summary file, we could merge the steps where proposal texts are created
vs. when they are submitted. Both could be done from a single script.
All it would have to know is

1. What is the release candidate commit ID.
2. Which canisters.
3. The ID of the neuron that will submit the proposals.

This could further be merged with list-new-commits, but we would only
want to do that if we trust list-new-commits to decide which canisters
to release.

At this point, I would say that the most tedious part of the process is
copying the proposal texts and URLs to various places. In particular,

1. The forum.
2. The #eng-nns-releases Slack channel.
3. The Calendar event.

I suspect 2 is the easiest to automate. It would just be an "extra"
thing that the submission script does after a successful submission.