refactor(cketh/ckerc20): mark older helper smart contracts as deprecated (#2747)
The helper smart contract supporting subaccount is the new preferred way
to do deposit, both of ETH and ERC-20. The 2 previous helper smart
contracts are therefore deprecated. Concretely,

1. The status (active or deprecated) of a helper smart contract is added
to the minter dashboard.
2. In case the address of the helper smart contract supporting
subaccount is set, the minter endpoints to retrieve the state of a
scraping (contract address and last scraped block number) for deposit of
ETH and ERC-20 will be that of the new helper smart contract supporting
subaccount.
3. The field `last_deposit_with_subaccount_scraped_block_number` was
removed. This is not a breaking change since it was never deployed in
production.