feat(IDX): don't rely on bazel cache for large test deps (#2752)
This changes the system tests to not assume that big dependencies like
ic-os images are present in the Bazel cache.

In particular, this adds an upload script (`upload_systest_dep.sh`)
which is run by the "local" (non-farm-host) test runner to ensure that
the images have been "stored". Currently the meaning of "stored" means
that the images are in the `bazel-remote` cache, though this is now an
implementation detail, encapsulated in the `upload_systest_dep` script.

This ensures that images are always ready when the tests are run, even
if the remote bazel cache was previously prune. This also gives us more
flexibility for implementing the Bazel cache differently. This also
simplifies the bazel build by not requiring both a URL & a sha256 file
for images; instead the file can be propagated directly.

---------

Co-authored-by: Bas van Dijk <bas@dfinity.org>