feat(orchestrator): CON-1440 Call `sync` and `fstrim` during node unassignment (#2720)
When a node is being unassigned, in the orchestrator upgrade loop, after
currently unlinking the files, we should sync the filesystem and issue
the TRIM command to the NVMe disks to ensure the blocks not used by the
filesystem are discarded.

As calling `fstrim` requires root access, this is implemented by
executing a separate script which calls `sync`, and then `fstrim` on the
replica's data directory.

In addition, we add metrics recording the number of failed `sync` &
`fstrim` calls, as well as the duration of the last call. These metrics
may be used to add some alerting in the future.

We adapt system tests for unassignment to assert the correct value of
these metrics.