fix(IDX): Ensure upload_systest_dep variables are defined
This defines all variables in `upload_systest_dep` are defined. Two
variables were not defined at the first use site due to a last minute
refactoring.

This PR also enables `set -u` which will ensure that bash fails if a
variable is used before being defined. This should have been enabled
when the changes were first introduced; this was an oversight.