chore(ci): Make `CI PR Only` mandatory for PRs (#2796)
This PR is a necessary step before making `CI PR Only` required /
mandatory on PRs.

Currently, when a workflow is marked as required for PR, it is also
enforced on the merge queue as well which can block the queue. However,
these jobs are only needed to gate the merge. Hence, we allow the
workflow to run on `merge_group` but each individual job is not run by
adding ` if: ${{ github.event_name != 'merge_group' }}`

---------

Co-authored-by: IDX GitHub Automation <IDX GitHub Automation>