test: create module for testing V2 update calls (#2803)
We want to upgrade the agent-rs version in the repo. The newer versions
do not support V2 update calls, which we currently use for system tests
to test the V2 call endpoint.

This PR creates a helper module for system tests to submit update calls
to the V2 call endpoint, and removes the usage of agent-rs for such
tests.