fix(IDX): correctly parse upload metrics (#2819)
Without a newline, `read` might fail to read the metrics. Previously,
this would fail the entire script.

Now, a newline is appended to the curl output, and the `read` call is
wrapped in an `if`. The curl output is also `tee`d to a file instead of
redirected directly, so that it can still be see on stderr.