refactor(CON): Rename dkg::Dealings to dkg::DataPayload (#2793)
This is a preparation for CON-1391.

This PR renames the DKG Payload type for non-summary blocks from
`Dealings` to `DkgDataPayload`.
The reason is that it no longer contains only dealings. There is already
a height stored in the block, and when CON-1391 will be implemented, the
payload will also contain transcripts. To avoid confusion, we are
renaming the Payload.