test(EXC-1808): Precompile universal canister (#2816)
EXC-1808

Add a build target which compiles the universal canister from Wasm to a
`Wasmtime::Module` and by default include this module to the compilation
cache for `ExecutionTest`s.

This will solve the problem of many tests compiling the universal
canister, which uses all system threads and generates CPU contention.