refactor(nervous_system_agent): Make CallCanisters implementation slightly more generic (#2724)
The `Request` trait in ic_nervous_system_clients is overly restrictive,
as it assumes that all canister endpoints take exactly one argument.
This is the case 99% of the time for our canisters, so it's nice to
have, but for those cases where a canister endpoint takes 0 arguments,
it's we need a more general option. This PR adds a trait that is
flexible w.r.t. the number of argument, and makes use of it in
ic_nervous_system_agent. There is also a blanket implementation, so
anything that implements the old trait also implements the new trait.

This adds a small amount of unfortunate complexity, but the benefit is
large, because it means that we can now use CallCanisters for types
which we could not implement `ic_nervous_system_clients::Request` for. I
demonstrate this by adding a NullRequest type (which was the real
motivation for this change), and using it to implement
`get_nervous_system_parameters` as a proof of concept