feat: Unify checkpointed_state and state in commit_and_certify (#2827)
When we create a checkpoint, we have two versions of `ReplicatedState`:
one that we loaded from disk and one came from execution (updated to the
new files). These two must be the same and we have a validation to raise
a critical error if they diverge.
This change eliminates `checkpointed_state`, so we have only one.