fix: make sure the root subnet is the first system subnet (#2809)
This PR ensures that the root subnet is the first system subnet.
Currently, it is not possible to use
`use_specified_ids_allocation_range` with multiple system subnets since
the specified ids allocation large is set for the first system subnet
while the root subnet is the last system subnet.