refactor: CON-1434 Deprecate `pb::TaggedNiDkgTranscript` and implement new deserialization (#2838)
As the `NiDkgId` (and therefore the `NiDkgTag`) is already part of the
`NiDkgTranscript`, serializing the `NiDkgTag` separately as part of
`TaggedNiDkgTranscript` is redundant.

This PR marks the `TaggedNiDkgTranscript` vectors for current and next
transcript in the summary payload as deprecated, and instead introduces
new vectors of `NiDkgTranscript`s.

Additionally, we extend the deserialization functions to consider both
fields. This needs to be done now already, such that we are prepared to
rollback once we start serializing the transcripts into the new fields.

This switch in serialization will be made once the change of this PR is
rolled out to every subnet.