fix(sns): SNS Gov should not accept an upgrade path from SNS-W that contains duplicate versions (#2839)
An upgrade path with duplicate versions from SNS-W would almost
certainly be erroneous, and likely to cause the SNS to enter an infinite
upgrade loop, which could be unrecoverable. We also may make the
assumption in various places that the upgrade steps will not have
duplicates, so it's good to enforce that.

[← Previous PR](https://github.com/dfinity/ic/pull/2836)