feat(ckbtc): Add account to discarded UTXOs (#2762)
Follow-up on #2674 to add the ledger account to a UTXO that was
discarded, either due to a value being too small to cover the KYT fee,
or being marked as tainted.

Remarks:

1. Add a new event type `DiscardedUtxo` that is used for both an ignored
and quarantined UTXO and that contains the ledger account.
2. All code paths were the ledger account was not needed to discard a
UTXO are marked as deprecated. They cannot be removed to be able to
replay past events.
3. Added metrics related to the number of discarded UTXOs without ledger
account since this should hopefully go to `0` and related to the number
of ignored/quarantined UTXOs.
4. `CI_OVERRIDE_DIDC_CHECK`: The Candid changes only affect the debug
endpoint `get_events`.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>