feat: [EXC-1767] Enable tail_call in wasmtime.  (#2826)
Enables the wasmtime feature
[tail_call](https://github.com/WebAssembly/tail-call/blob/master/proposals/tail-call/Overview.md).

Benchmarks show that using `return_call` instead of `call` can improve
performance significantly, but the improvement depends on the depth of
the call stack. In short, with reasonable assumptions about call depth,
the improvement is about 1.5x, so the cost for `return_call` and
`return_call_indirect` are adjusted downward to incentivize the use of
this optimization by the toolchains.

---------

Co-authored-by: Michael Weigelt <michael.weigelt@dfinity.com>