refactor(sns): Reset cached upgrade steps upon detecting inconsistencies (#2802)
This PR refactors the SNS Governance to reset cached upgrade steps
earlier, as soon as an inconsistency or missing data is detected. This
simplifies the implementation of a few periodic functions since they can
rely on the data being available.

The behavior is slightly affected, as the first upgrade journal entry
for a newly deployed SNS is expected to be a cache _reset_ rather than a
cache refresh.

---------

Co-authored-by: Andre Popovitch <andre@popovit.ch>