feat(IC-1579): TLA instrumentation for `merge_neuron` (#2341)
Adds a TLA model of `merge_neurons` and links it to the code.

Since the only Rust integration tests that test merge_neuron are
proptests, which generate a ton of traces, introduced a limit to check
only some prefix (currently, first 30) of the state pairs from the
collected traces. Also made the Apalache checks a bit more parallel, by
using channels to signal completion as soon as a single thread is
available instead of waiting for the batch.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>