test: CON-1440 Assert that a subnet state exists on the node before it is deleted (#2916)
In the `node_reassignment_test` we want to verify that a node's state is
correctly delete during unassignment. In
https://github.com/dfinity/ic/pull/2720 we therefore added an assertion
checking that the state of unassigned nodes is empty. Technically, this
check would also succeed if the state never existed in the first place.

For that reason, this PR adds an assertion to make sure that the state
exists before unassignment, and no longer exists afterwards.
Additionally, we extend the assertion to not only check the node state,
but also the orchestrator's local CUP file.