feat(nns): Added prune_some_following. (#2471)
Remove following when a neuron has not refreshed in more than 7 months.

The one exception is ManageNeuron; following on that topic is retained
forever.

In a later change (soon), this will be called periodically in the
background of the NNS governance canister. For now, this does not result
in any externally visible behavior changes to any canisters.

# References

[👈 Previous PR][prev] | [Next PR 👉][next]

[prev]: https://github.com/dfinity/ic/pull/2470
[next]: https://github.com/dfinity/ic/pull/2392