fix(IDX): Add default directory for lint script (#2937)
This defaults the working directory of the lint script to the root of
the git repo, in case `CI_PROJECT_DIR` is not set.

Before this, the variable `CI_PROJECT_DIR` would be read, which is only
set on CI. By adding a default, the script can be run locally.