feat(IDX): Force static link of lmdb
This ensures that lmdb is linked in statically in the Bazel build, and
that the specified archive is used.

Before this, lmdb was linked statically somewhat by chance, and only at
the end of the build. Now lmdb is linked in when the `lmdb-rkv-sys`
crate is built and doesn't need to be linked when final executables are
built.