perf: Copy outside write lock (#2963)
In the state manager's commit_and_certify function, it now does the
copy_state step before taking the write lock for the list of states.

Several API calls to the state manager rely on reading from the list of
states, and they are blocked otherwise.

This fixes a regression introduced by commit

[e3ee007](https://github.com/dfinity/ic/commit/e3ee0072c1ea003c489b59fa10df2610c60fe2e6)