refactor(nns): Move recompute_tally to inside of cast_vote_and_cascade_follow (#2977)
This removes a difficult edge case of trying to determine when to
re-tally. We know that any votes cast in cast_vote_and_cascade_follow
are either when a proposal is created, or when register_vote is called,
which makes sure the vote cast is cast while the proposal voting is
still open.

This helps later when votes can be processed across multiple messages.